requirements.txt
python-telegram-bot==20.7
ccxt==4.1.77
web3==6.11.3
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
aiohttp==3.9.1
websockets==12.0